+++ Cargo.toml
@@ -11,7 +11,7 @@ license = "LGPL-3.0 OR Apache-2.0"
 
 [dependencies]
 libc = "0.2.119"
-open-coroutine-core = { version = "0.4.0", path = "open-coroutine-core" }
+open-coroutine-core = { version = "0.4.0", path = "open-coroutine-core", features = ["full"] }
 open-coroutine-macros = { version = "0.1.0", path = "open-coroutine-macros" }
 
 [workspace]
