+++ src/tests/mod.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+mod sleep;
+
+mod socket;
+
+#[test]
+fn test_link() {
+    init();
+}
